+++ FloorPlanGenerator/lib/cvHelper.h
@@ -23,7 +23,7 @@ public:
      * @param[in] ptsY y axis values
      * @return None
     */
-    static void showLayout(const std::vector<int> &ptsX, const std::vector<int> &ptsY);
+    static void showLayout(const std::vector<int16_t> &ptsX, const std::vector<int16_t> &ptsY);
 };
 
 #endif //OPENCV_HELPER
