+++ FloorPlanGenerator/lib/generate.h
@@ -25,7 +25,7 @@ public:
         @param[in] rooms vector containg all rooms informations, such as minimum and maximum sizes
         @return vector of vector of vector of layout combination. result[a][b][c] = d, a -> room size id, b -> permutation id, d -> layout points
     */
-    static std::vector<std::vector<std::vector<int>>> SizeLoop(const std::vector<RoomConfig>& rooms);
+    static std::vector<std::vector<std::vector<int16_t>>> SizeLoop(const std::vector<RoomConfig>& rooms);
 
     /*!
         @brief Iterate over every possible connection between the given rooms 
@@ -36,7 +36,7 @@ public:
         @param[in] NConn Number of possible connections
         @return vector with layout points for every successful connection (n*4 int per layout)
     */
-    static std::vector<int> ConnLoop(const std::vector<int>& order, const int *sizeH, const int *sizeW, const int n, const int NConn);
+    static std::vector<int16_t> ConnLoop(const std::vector<int>& order, const int16_t *sizeH, const int16_t *sizeW, const int n, const int NConn);
 
 
     /*!
@@ -46,7 +46,7 @@ public:
         @param[in] n     number of rooms
         @return  vector of vector of layout combination. result[a][b] = c, a -> permutation id, c -> layout points
     */
-    static std::vector<std::vector<int>> roomPerm(const int *sizeH, const int *sizeW, const int n);
+    static std::vector<std::vector<int16_t>> roomPerm(const int16_t *sizeH, const int16_t *sizeW, const int n);
 
 };
